+++ src/include/uimusicq.h
@@ -149,7 +149,7 @@ void  uimusicqTruncateQueue (uimusicq_t *uimusicq, int mqidx, nlistidx_t idx);
 void  uimusicqPlay (uimusicq_t *uimusicq, int mqidx, dbidx_t dbidx);
 void  uimusicqQueue (uimusicq_t *uimusicq, int mqidx, dbidx_t dbidx);
 void  uimusicqSetPeerFlag (uimusicq_t *uimusicq, bool val);
-bool  uimusicqSaveListCallback (void *udata, long dbidx);
+bool  uimusicqSaveListCallback (void *udata, int32_t dbidx);
 
 #endif /* INC_UIMUSICQ_H */
